DelegatingAuroraConnection: lockless implementation when auto-commit
This change attempts to remove the locking when detecting the delegate connection if auto-commit is enabled.
When auto-commit is enabled this also will reduce the Pair object creation.  However this does require double-checking the auto commit volatile when auto-commit is disabled.
Since I believe auto-commit is used most of the time, this should be a net positive.  I also believe this may make future changes easier.